Initial files of the project
Add: all the initial files needed for the project

Currently the app can draw with the standard black color, an gray option and a 'random color' option. It also has an 'eraser', everything working.

Need to add: only draw when holding the the click button. Currently draws everytime you hover the mouse over the 'board'